friend_request_app/requirements.txt
aiosqlite==0.17.0
allure-pytest==2.13.2
allure-python-commons==2.13.2
altgraph==0.17.3
anyio==3.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
autopep8==2.0.2
awscli==1.33.26
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
botocore==1.34.144
build==1.2.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.6
comm==0.1.3
contourpy==1.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
distlib==0.3.7
distro==1.7.0
Django==4.2.9
django-allauth==0.61.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docopt==0.6.2
docutils==0.16
drf-yasg==1.21.7
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.96.0
fastapi-mail==1.2.8
fastjsonschema==2.18.0
filelock==3.12.4
fonttools==4.42.1
fqdn==1.5.1
greenlet==2.0.2
h11==0.14.0
httplib2==0.20.2
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.3
iso8601==1.1.0
isoduration==20.11.0
itypes==1.2.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
LinkChecker==10.4.0
macaroonbakery==1.3.1
Markdown==3.3.6
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.1
more-itertools==8.10.0
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.9.2
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.0
openapi-codec==1.3.2
opencv-python==4.10.0.82
openpyxl==3.1.2
ordered-set==4.1.0
outcome==1.2.0
overrides==7.4.0
packaging==24.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pbr==5.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip-install-test==0.5
pip-tools==7.4.1
pipdeptree==2.23.1
pipenv==2023.10.20
pipreqs==0.5.0
platformdirs==3.11.0
pluggy==1.3.0
postgres==4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==3.12.4
psutil==5.9.4
psycopg2==2.9.5
psycopg2-binary==2.9.9
psycopg2-pool==1.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydantic==1.10.8
pydentic==0.0.1.dev3
Pygments==2.11.2
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypika-tortoise==0.1.6
pyproject_hooks==1.1.0
pyRFC3339==1.1
PySocks==1.7.1
pytest==7.4.2
pytest-html==4.0.2
pytest-metadata==3.0.0
pytest-reporter==0.4.0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-stdnum==1.18
python3-openid==3.2.0
pytz==2023.3.post1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.0.2
randomcolor==0.4.4.6
ratelimiter==1.2.0.post0
redis==4.4.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.7.2
ruamel.yaml==0.17.21
s3transfer==0.10.2
schedule==1.2.0
scikit-learn==1.3.1
scipy==1.11.2
SecretStorage==3.3.1
selenium==4.13.0
Send2Trash==1.8.2
servicenow==2.1.0
servicenow-api==0.13.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
SOAPpy==0.12.22
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.7
sqlparse==0.4.4
ssh-import-id==5.11
stack-data==0.6.2
starlette==0.27.0
style==1.1.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tortoise-orm==0.19.3
tqdm==4.65.1
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.9.0
update==0.0.1
urllib3==2.2.2
uri-template==1.3.0
uritemplate==4.1.1
uvicorn==0.22.0
virtualenv==20.24.6
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
wlc==1.2
wsproto==1.2.0
wstools==0.4.8
xdg==5
xlrd==2.0.1
yarg==0.1.9
zipp==1.0.0
